Updated data config. Implemented related ui for top rated movies displaying.

+++ domain/src/main/java/com/vlohachov/domain/model/Movie.kt
@@ -9,5 +9,5 @@ data class Movie(
     val genreIds: List<Long>,
     val isAdult: Boolean,
     val voteCount: Long,
-    val voteAverage: Int,
+    val voteAverage: Float,
 )
